+++ docs/changelog.md
@@ -12,6 +12,7 @@ category: Administration
 ### [Unreleased]
 
 * docs: add frequently asked questions to online documentation
+* feat: support estimate fields that are select lists or radio buttons (#120)
 
 ### [20.12.1] - 2020-12-07